fastapp/requirements.txt
fastapi
uvicorn[standard]
websocket-client
requests
python-multipart

Pillow
